requirements.txt
pytorch
pytorch-geometric
networkx
numpy
scipy